OneDrive/Desktop/hospital-management/hospital-management/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
 
